+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/preprocessor/utils/attributes-to-values.test.ts
@@ -1,0 +1,65 @@
+import { describe, it, expect } from "vitest"
+import { attrubuteValuesToJSValue } from "./attributes-to-values"
+import type { AttributeValue } from "../types"
+
+describe("attributesToValues", () => {
+	it("stringifies a single text value", () => {
+		const originalCode = "hello"
+
+		const attributeValues: AttributeValue[] = [
+			{
+				start: 0,
+				end: 5,
+				type: "Text",
+				raw: "hello",
+				data: "hello",
+			},
+		]
+
+		const result = attrubuteValuesToJSValue(attributeValues, originalCode)
+		expect(result).toBe("`hello`")
+	})
+
+	it("stringifies a single mustache tag", () => {
+		const originalCode = "{hello}"
+		const attributeValues: AttributeValue[] = [
+			{
+				start: 0,
+				end: 7,
+				type: "MustacheTag",
+				expression: {
+					start: 1,
+					end: 6,
+				},
+			},
+		]
+
+		const result = attrubuteValuesToJSValue(attributeValues, originalCode)
+		expect(result).toBe("`${hello}`")
+	})
+
+	it("stringifies an interweaved text and mustache tag", () => {
+		const originalCode = "hello {world}"
+		const attributeValues: AttributeValue[] = [
+			{
+				start: 0,
+				end: 5,
+				type: "Text",
+				raw: "hello ",
+				data: "hello ",
+			},
+			{
+				start: 6,
+				end: 13,
+				type: "MustacheTag",
+				expression: {
+					start: 7,
+					end: 12,
+				},
+			},
+		]
+
+		const result = attrubuteValuesToJSValue(attributeValues, originalCode)
+		expect(result).toBe("`hello ${world}`")
+	})
+})
